AI_chat/requirements.txt
asgiref==3.8.1
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
distlib==0.3.8
distro-info==0.23+ubuntu1.1
dj-database-url==2.2.0
Django==4.2.14
django-cors-headers==4.4.0
django-filter==24.3
django-heroku==0.3.1
django-taggit==6.0.0
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==23.0.0
idna==2.8
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyGObject==3.36.0
PyJWT==2.9.0
python-apt==2.0.1+ubuntu0.20.4.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
sqlparse==0.5.1
typing_extensions==4.12.2
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.26.3
whitenoise==6.7.0
